Add invisible `<button>` for keyboard access
Since the button doesn't respond to pointer events, it can only be
activated with the keyboard.

I'm hoping that screen readers will be able to activate the button the
same way the keyboard does, but I'm not 100% sure that actually works.
It would be better if we could let the button respond to pointer events,
but that would require more invasive changes and hopefully won't be
necessary.